task_6/server/requirements.txt
Flask==2.2.5
lightgbm==4.0.0
numpy==1.26.0
pandas==2.2.1
requests==2.31.0
python-dotenv==1.0.0
